Move logic out of dump manager
