+++ listenbrainz/db/dump.py
@@ -43,6 +43,11 @@ from listenbrainz.utils import create_path, log_ioerrors
 
 from listenbrainz import config
 
+
+MAIN_FTP_SERVER_URL = "ftp.eu.metabrainz.org"
+FULLEXPORT_MAX_AGE = 16  # days
+INCREMENTAL_MAX_AGE = 26  # hours
+
 # this dict contains the tables dumped in public dump as keys
 # and a tuple of columns that should be dumped as values
 PUBLIC_TABLES = {
@@ -677,6 +682,74 @@ def _update_sequences():
     # data_dump_id_seq
     current_app.logger.info('Updating data_dump_id_seq...')
     _update_sequence('data_dump_id_seq', 'data_dump')
+
+
+def _fetch_latest_file_info_from_ftp_dir(server, dir):
+    """
+        Given a FTP server and dir, fetch the latst dump file info, parse it and
+        return a tuple containing (dump_id, datetime_of_dump_file).
+    """
+
+    line = ""
+
+    def add_line(l):
+        nonlocal line
+        l = l.strip()
+        if l:
+            line = l
+
+    ftp = FTP(server)
+    ftp.login()
+    ftp.cwd(dir)
+    ftp.retrlines('LIST', add_line)
+    _, _, id, d, t, _ = line[56:].strip().split("-")
+
+    return int(id), datetime.strptime(d + t, "%Y%m%d%H%M%S")
+
+
+def check_ftp_dump_ages():
+    """
+        Fetch the FTP dir listing of the full and incremental dumps and check their ages. Send mail
+        to the observability list in case the dumps are too old.
+    """
+
+    msg = ""
+    try:
+        id, dt = _fetch_latest_file_info_from_ftp_dir(
+            MAIN_FTP_SERVER_URL, '/pub/musicbrainz/listenbrainz/fullexport')
+        age = datetime.now() - dt
+        if age > timedelta(days=FULLEXPORT_MAX_AGE):
+            msg = "Full dump %d is more than %d days old: %s\n" % (
+                id, FULLEXPORT_MAX_AGE, str(age))
+            print(msg, end="")
+        else:
+            print("Full dump %s is %s old, good!" % (id, str(age)))
+    except Exception as err:
+        msg = "Cannot fetch full dump age: %s" % str(err)
+
+    try:
+        id, dt = _fetch_latest_file_info_from_ftp_dir(
+            MAIN_FTP_SERVER_URL, '/pub/musicbrainz/listenbrainz/incremental')
+        age = datetime.now() - dt
+        if age > timedelta(hours=INCREMENTAL_MAX_AGE):
+            msg = "Incremental dump %s is more than %s hours old: %s\n" % (
+                id, INCREMENTAL_MAX_AGE, str(age))
+            print(msg, end="")
+        else:
+            print("Incremental dump %s is %s old, good!" % (id, str(age)))
+    except Exception as err:
+        msg = "Cannot fetch full dump age: %s" % str(err)
+
+    app = create_app()
+    with app.app_context():
+        if not current_app.config['TESTING'] and msg:
+            send_mail(
+                subject="ListenBrainz outdated dumps!",
+                text=render_template('emails/data_dump_outdated.txt', msg=msg),
+                recipients=['listenbrainz-exceptions@metabrainz.org'],
+                from_name='ListenBrainz',
+                from_addr='noreply@'+current_app.config['MAIL_FROM_DOMAIN']
+            )
 
 
 class SchemaMismatchException(Exception):
